+++ src/http/arraybuffer.request.client.ts
@@ -1,21 +1,6 @@
-import { Resource } from "@companieshouse/api-sdk-node";
-import { File } from "private-api-sdk-node/dist/services/file-transfer/types";
 import axios, { AxiosRequestConfig, AxiosResponse } from "axios";
 import { CHS_INTERNAL_API_KEY, INTERNAL_API_URL } from "../config";
-import Mapping from "@companieshouse/api-sdk-node/dist/mapping/mapping";
-import { ApiErrorResponse } from "@companieshouse/api-sdk-node/dist/services/resource";
 
-
-export const createLocalAPIClient = (): LocalAPIClient => {
-    return new LocalAPIClient(new ArrayBufferRequestClient());
-};
-
-export default class LocalAPIClient {
-    readonly arrayBufferRequestClient: ArrayBufferRequestClient;
-    constructor(arrayBufferRequestClient: ArrayBufferRequestClient){
-        this.arrayBufferRequestClient = arrayBufferRequestClient;
-    }
-}
 
 /**
  * TODO: find a better way of getting this functionality into the ecosystem,
@@ -25,56 +10,13 @@ export default class LocalAPIClient {
 export class ArrayBufferRequestClient {
 
     /**
-     * TODO: Move this code to a better fit for it.
-     * This code has been add here due to it being only required by ImageRender's render
-     * and other factor.
-     * GET method to get the rendered file as PDF
-     * @param fileId
-     * @returns
-     */
-    async rawGetRenderedPDF(
-        fileId: string
-    ): Promise<Resource<File> | ApiErrorResponse> {
-        const mimeType = "application/pdf";
-        const fileNameTemplate = `${fileId}`;
-        const extension = "pdf";
-        const resp = await this.rawGetArrayBuffer(`/validate/render/${fileId}`);
-
-        if (resp.status !== 200) {
-            return {
-                httpStatusCode: resp.status,
-                errors: []
-            };
-        }
-
-        const headers = Mapping.camelCaseKeys(resp.headers);
-        const contentType = headers?.["contentType"];
-        if (contentType === undefined || contentType === null || (contentType !== "application/pdf")) {
-            throw new Error("Content type is not application/pdf");
-        }
-
-        const pdf: File = {
-            fileName: fileNameTemplate,
-            body: resp.body,
-            mimeType: mimeType,
-            size: resp.headers.byteLength,
-            extension: extension
-        };
-
-        return {
-            httpStatusCode: resp.status,
-            resource: pdf
-        };
-    }
-
-    /**
     * TODO: Move this code to a better fit for it.
     * This code has been add here due to it being only required by ImageRender's render
     * and other factor
      * @param url path to the endpoint
      * @returns
      */
-    private async rawGetArrayBuffer(url: string) {
+    async rawGetArrayBuffer(url: string) {
         try {
             const options: AxiosRequestConfig = {
                 method: "GET",
